ETSWebServices: Restructuration part 1
In an attempt to remove some responsibilities from ETSMobile, place
all SOAP related webservices in a single project containing the
required modules. They are as follow:

- etsmtl-webservices-core: Base library/module containing required
classes to modules that depend on it. (In order to avoid code
duplication)
- etsmtl-webservices-directory: Library used in order to communicate
with ETS's employee directory
- etsmtl-webservices-signets: Library used to communicate with
SIGNETS in order to give a student academic related stuff
(Results, schedule, etc...)

All of them have been rendered from easywsdl.com for easy
convertion from WSDL to Java.

The structure might not be the best, but at least it is the first
step in a cleaner "environment" one might say.